Do refactor packages and add throughput log on single process
