+++ datamimic_ce/tasks/generate_task.py
@@ -7,9 +7,6 @@
 import copy
 import math
 import shutil
-import time
-from contextlib import contextmanager
-from typing import Literal
 
 import dill  # type: ignore
 import ray
@@ -18,9 +15,6 @@ from datamimic_ce.clients.database_client import DatabaseClient
 from datamimic_ce.contexts.context import Context
 from datamimic_ce.contexts.geniter_context import GenIterContext
 from datamimic_ce.contexts.setup_context import SetupContext
-from datamimic_ce.data_sources.data_source_pagination import DataSourcePagination
-from datamimic_ce.data_sources.data_source_util import DataSourceUtil
-from datamimic_ce.exporters.exporter_state_manager import ExporterStateManager
 from datamimic_ce.logger import logger
 from datamimic_ce.statements.composite_statement import CompositeStatement
 from datamimic_ce.statements.generate_statement import GenerateStatement
@@ -28,173 +22,7 @@ from datamimic_ce.statements.key_statement import KeyStatement
 from datamimic_ce.statements.statement import Statement
 from datamimic_ce.tasks.task import CommonSubTask
 from datamimic_ce.utils.base_class_factory_util import BaseClassFactoryUtil
-
-
-def generate_product_by_page_in_single_process(
-        context: SetupContext | GenIterContext, stmt: GenerateStatement, page_start: int, page_end: int, worker_id: int
-) -> dict[str, list]:
-    """
-    (IMPORTANT: Only to be used as Ray multiprocessing function)
-    This function is used to generate data for a single process, includes steps:
-    1. Build sub-tasks in GenIterStatement
-    2. Load data source (if any)
-    3. Modify/Generate data by executing sub-tasks
-
-    :return: Dictionary with generated products.
-    """
-    root_context: SetupContext = context.root
-
-    # Determine number of data to be processed
-    processed_data_count = page_end - page_start
-    pagination = DataSourcePagination(skip=page_start, limit=processed_data_count)
-
-    # Determined page of data source to load
-    # If distribution is random, load all data before shuffle, which means no pagination (or, pagination=None)
-    # If distribution is not random, load data by pagination
-    is_random_distribution = stmt.distribution in ("random", None)
-    if is_random_distribution:
-        # Use task_id as seed for random distribution
-        # Don't use pagination for random distribution to load all data before shuffle
-        load_start_idx = None
-        load_end_idx = None
-        load_pagination: DataSourcePagination | None = None
-    else:
-        load_start_idx = page_start
-        load_end_idx = page_end
-        load_pagination = pagination
-
-    # Extract converter list for post-processing
-    task_util_cls = root_context.class_factory_util.get_task_util_cls()
-    converter_list = task_util_cls.create_converter_list(context, stmt.converter)
-
-    # 1: Build sub-tasks in GenIterStatement
-    tasks = [
-        task_util_cls.get_task_by_statement(root_context, child_stmt, pagination) for child_stmt in stmt.sub_statements
-    ]
-
-    # 2: Load data source from file, database, memory, Kafka, etc.
-    source_str = stmt.source
-    source_scripted = (
-        stmt.source_script if stmt.source_script is not None else bool(root_context.default_source_scripted)
-    )
-    separator = stmt.separator or root_context.default_separator
-
-    source_data, build_from_source = context.root.class_factory_util.get_task_util_cls().gen_task_load_data_from_source(
-        context,
-        stmt,
-        source_str,
-        separator,
-        source_scripted,
-        processed_data_count,
-        load_start_idx,
-        load_end_idx,
-        load_pagination,
-    )
-
-    # Shuffle source data if distribution is random
-    if is_random_distribution:
-        seed = root_context.get_distribution_seed()
-        # Use original pagination for shuffling
-        source_data = DataSourceUtil.get_shuffled_data_with_cyclic(source_data, pagination, stmt.cyclic, seed)
-
-    # Store temp result
-    product_holder: dict[str, list] = {}
-    result = []
-
-    # 3: Modify/Generate data by executing sub-tasks
-    for idx in range(processed_data_count):
-        # Create sub-context for each product record creation
-        ctx = GenIterContext(context, stmt.name)
-        # Get current worker_id from outermost gen_stmt
-        ctx.worker_id = worker_id
-
-        # Set current product to the product from data source if building from datasource
-        if build_from_source:
-            if idx >= len(source_data):
-                break
-            ctx.current_product = copy.deepcopy(source_data[idx])
-
-        try:
-            # Start executing sub-tasks
-            from datamimic_ce.tasks.condition_task import ConditionTask
-
-            for task in tasks:
-                # Collect product from sub-generate task and add into product_holder
-                if isinstance(task, GenerateTask | ConditionTask):
-                    # Execute sub generate task
-                    sub_gen_result = task.execute(ctx)
-                    if sub_gen_result:
-                        for key, value in sub_gen_result.items():
-                            # Store product for later export
-                            product_holder[key] = product_holder.get(key, []) + value
-                            # Store temp product in context for later evaluate
-                            inner_generate_key = key.split("|", 1)[-1].strip()
-                            ctx.current_variables[inner_generate_key] = value
-                else:
-                    task.execute(ctx)
-
-            # Post-process product by applying converters
-            for converter in converter_list:
-                ctx.current_product = converter.convert(ctx.current_product)
-
-            # Lazily evaluate source script after executing sub-tasks
-            if source_scripted:
-                # Evaluate python expression in source
-                prefix = stmt.variable_prefix or root_context.default_variable_prefix
-                suffix = stmt.variable_suffix or root_context.default_variable_suffix
-                ctx.current_product = task_util_cls.evaluate_file_script_template(
-                    ctx=ctx, datas=ctx.current_product, prefix=prefix, suffix=suffix
-                )
-
-            result.append(ctx.current_product)
-        except StopIteration:
-            # Stop generating data if one of datasource reach the end
-            logger.info(
-                f"Data generator sub-task {task.__class__.__name__} '{task.statement.name}' has already reached the end"
-            )
-            break
-
-    # 4. Return product for later export
-    product_holder[stmt.full_name] = result
-    return product_holder
-
-
-@contextmanager
-def gen_timer(process: Literal["generate", "export", "process"], report_logging: bool, product_name: str):
-    """
-    Timer for generate and export process.
-
-    :param process: Type of process ('generate', 'consume', 'process').
-    :param report_logging: Whether to log the timing information.
-    :param product_name: Name of the product being processed.
-    :return: Context manager.
-    """
-    timer_result: dict = {}
-    # Ignore timer if report_logging is False
-    if not report_logging:
-        yield timer_result
-        return
-    start_time = time.time()
-    try:
-        yield timer_result
-    finally:
-        records_count = timer_result.get("records_count", 0)
-        end_time = time.time()
-        elapsed_time = end_time - start_time
-        timer_result["elapsed_time"] = elapsed_time
-        match process:
-            case "generate":
-                process_name = "Generating"
-            case "export":
-                process_name = "Exporting"
-            case _:
-                process_name = "Generating and exporting"
-        logger.info(
-            f"{process_name} {records_count:,} records '{product_name}' took {round(elapsed_time, 5)} seconds "
-            f"({int(records_count / elapsed_time):,} records/second)"
-            if elapsed_time > 0
-            else "N/A records/second"
-        )
+from datamimic_ce.utils.logging_util import gen_timer
 
 
 class GenerateTask(CommonSubTask):
@@ -304,7 +132,7 @@ class GenerateTask(CommonSubTask):
                 GenerateTask._scan_data_source(ctx, child_stmt)
 
     @staticmethod
-    def determine_num_workers(context: GenIterContext | SetupContext, stmt: GenerateStatement) -> int:
+    def _determine_num_workers(context: GenIterContext | SetupContext, stmt: GenerateStatement) -> int:
         """
         Determine number of Ray workers for multiprocessing. Default to 1 if not specified.
         Do not apply multiprocessing (return 1) if:
@@ -345,93 +173,97 @@ class GenerateTask(CommonSubTask):
         :param context: Context instance.
         :return: Generated product data or None.
         """
-        try:
-            # Pre-execute sub-tasks before generating any data
-            self.pre_execute(context)
+        with gen_timer("process", context.root.report_logging, self.statement.full_name) as timer_result:
+            try:
+                # Pre-execute sub-tasks before generating any data
+                self.pre_execute(context)
 
-            # Determine count of generate process
-            count = self._determine_count(context)
+                # Determine count of generate process
+                count = self._determine_count(context)
+                timer_result["records_count"] = count
 
-            # Early return if count is 0
-            if count == 0:
-                return {self.statement.full_name: []}
+                # Early return if count is 0
+                if count == 0:
+                    return {self.statement.full_name: []}
 
-            # Calculate page size for processing by page
-            page_size = self._calculate_default_page_size(count)
+                # Calculate page size for processing by page
+                page_size = self._calculate_default_page_size(count)
 
-            # Determine number of Ray workers for multiprocessing
-            num_workers = self.determine_num_workers(context, self.statement)
+                # Determine number of Ray workers for multiprocessing
+                num_workers = self._determine_num_workers(context, self.statement)
 
-            # Execute generate task by page in multiprocessing
-            if isinstance(context, SetupContext) and num_workers > 1:
-                # Serialize context for Ray multiprocessing
-                copied_context = copy.deepcopy(context)
-                ns_funcs = {k: v for k, v in copied_context.root.namespace.items() if callable(v)}
-                for func in ns_funcs:
-                    copied_context.root.namespace.pop(func)
-                copied_context.root.namespace_functions = dill.dumps(ns_funcs)
-                copied_context.root.generators = dill.dumps(copied_context.root.generators)
+                # Execute generate task by page in multiprocessing
+                if isinstance(context, SetupContext) and num_workers > 1:
+                    # Serialize context for Ray multiprocessing
+                    copied_context = copy.deepcopy(context)
+                    ns_funcs = {k: v for k, v in copied_context.root.namespace.items() if callable(v)}
+                    for func in ns_funcs:
+                        copied_context.root.namespace.pop(func)
+                    copied_context.root.namespace_functions = dill.dumps(ns_funcs)
+                    copied_context.root.generators = dill.dumps(copied_context.root.generators)
 
-                # Determine chunk data indices based on chunk size and required data count
-                # then populate to workers, such as (0, 1000), (1000, 2000), etc.
-                chunk_size = math.ceil(count / num_workers)
-                chunks = [(i * chunk_size, min((i + 1) * chunk_size, count)) for i in range(num_workers)]
+                    # Determine chunk data indices based on chunk size and required data count
+                    # then populate to workers, such as (0, 1000), (1000, 2000), etc.
+                    chunk_size = math.ceil(count / num_workers)
+                    chunks = [(i * chunk_size, min((i + 1) * chunk_size, count)) for i in range(num_workers)]
 
-                # Execute generate task using Ray
-                futures = [
-                    RayGenerateWorker.generate_and_export_data_by_chunk_mp.options(enable_task_events=False).remote(
-                        copied_context,
-                        self._statement,
-                        worker_id,
-                        chunk_start,
-                        chunk_end, page_size)
-                    for
-                    worker_id, (chunk_start, chunk_end) in enumerate(chunks, 1)]
-                # Gather result from Ray workers
-                ray_result = ray.get(futures)
+                    # Execute generate task using Ray
+                    from datamimic_ce.workers.ray_generate_worker import RayGenerateWorker
+                    futures = [
+                        RayGenerateWorker.generate_and_export_data_by_chunk_mp.options(enable_task_events=False).remote(
+                            copied_context,
+                            self._statement,
+                            worker_id,
+                            chunk_start,
+                            chunk_end, page_size)
+                        for
+                        worker_id, (chunk_start, chunk_end) in enumerate(chunks, 1)]
+                    # Gather result from Ray workers
+                    ray_result = ray.get(futures)
 
-                # Merge result from all workers by product name
-                merged_result: dict[str, list] = {}
-                for result in ray_result:
-                    for product_name, product_data_list in result.items():
-                        merged_result[product_name] = merged_result.get(product_name, []) + product_data_list
-            # Execute generate task by page in single process
-            else:
-                # If inner gen_stmt, pass worker_id from outermost gen_stmt to inner gen_stmt
-                worker_id = context.worker_id if isinstance(context, GenIterContext) else 1
-                merged_result = GenerateWorker.generate_and_export_data_by_chunk(
-                    context, self._statement, worker_id, 0, count, page_size
-                )
+                    # Merge result from all workers by product name
+                    merged_result: dict[str, list] = {}
+                    for result in ray_result:
+                        for product_name, product_data_list in result.items():
+                            merged_result[product_name] = merged_result.get(product_name, []) + product_data_list
+                # Execute generate task by page in single process
+                else:
+                    # If inner gen_stmt, pass worker_id from outermost gen_stmt to inner gen_stmt
+                    worker_id = context.worker_id if isinstance(context, GenIterContext) else 1
+                    from datamimic_ce.workers.generate_worker import GenerateWorker
+                    merged_result = GenerateWorker.generate_and_export_data_by_chunk(
+                        context, self._statement, worker_id, 0, count, page_size
+                    )
 
-            # At the end of OUTERMOST gen_stmt:
-            # - export gathered product with LAZY exporters, such as TestResultExporter, MemstoreExporter,...
-            # - gather and upload ARTIFACT exporters, such as TXT, JSON,... then clean temp directory
-            if isinstance(context, SetupContext):
-                # Determine lazy exporters
-                lazy_exporters = []
-                if context.test_mode:
-                    lazy_exporters.append(context.root.test_result_exporter)
-                for exporter_str in self.statement.targets:
-                    if context.memstore_manager.contain(exporter_str):
-                        lazy_exporters.append(context.memstore_manager.get_memstore(exporter_str))
-                # Export gathered product with lazy exporters
-                for exporter in lazy_exporters:
-                    for product_name, product_records in merged_result.items():
-                        exporter.consume((product_name, product_records))
+                # At the end of OUTERMOST gen_stmt:
+                # - export gathered product with LAZY exporters, such as TestResultExporter, MemstoreExporter,...
+                # - gather and upload ARTIFACT exporters, such as TXT, JSON,... then clean temp directory
+                if isinstance(context, SetupContext):
+                    # Determine lazy exporters
+                    lazy_exporters = []
+                    if context.test_mode:
+                        lazy_exporters.append(context.root.test_result_exporter)
+                    for exporter_str in self.statement.targets:
+                        if context.memstore_manager.contain(exporter_str):
+                            lazy_exporters.append(context.memstore_manager.get_memstore(exporter_str))
+                    # Export gathered product with lazy exporters
+                    for exporter in lazy_exporters:
+                        for product_name, product_records in merged_result.items():
+                            exporter.consume((product_name, product_records))
 
-                # # Finalize chunks files (writing end of file)
-                self.finalize_temp_files_chunks(context, self.statement)
+                    # # Finalize chunks files (writing end of file)
+                    self.finalize_temp_files_chunks(context, self.statement)
 
-                # Upload ARTIFACT files at the end of ALL chunks processes
-                self.export_artifact_files(context, self.statement)
+                    # Upload ARTIFACT files at the end of ALL chunks processes
+                    self.export_artifact_files(context, self.statement)
 
-            return merged_result
+                return merged_result
 
-        finally:
-            # Clean temp directory on outermost gen_stmt
-            if isinstance(context, SetupContext):
-                for temp_dir in context.descriptor_dir.glob(f"temp_result_{context.task_id}*"):
-                    shutil.rmtree(temp_dir)
+            finally:
+                # Clean temp directory on outermost gen_stmt
+                if isinstance(context, SetupContext):
+                    for temp_dir in context.descriptor_dir.glob(f"temp_result_{context.task_id}*"):
+                        shutil.rmtree(temp_dir)
 
     @staticmethod
     def finalize_temp_files_chunks(context: SetupContext, stmt: GenerateStatement):
@@ -439,7 +271,7 @@ class GenerateTask(CommonSubTask):
         Finalize temp files chunks (writing end of file).
         """
         # Determine number of Ray workers, used to determined chunk directory name
-        num_workers = GenerateTask.determine_num_workers(context, stmt)
+        num_workers = GenerateTask._determine_num_workers(context, stmt)
 
         # Get ARTIFACT exporters from statement
         exporter_list = context.root.class_factory_util.get_exporter_util().get_all_exporter(context, stmt,
@@ -490,120 +322,3 @@ class GenerateTask(CommonSubTask):
         ]
         for task in pre_tasks:
             task.pre_execute(context)
-
-
-class GenerateWorker:
-    """
-    Worker class for generating and exporting data by page in single process.
-    """
-
-    @staticmethod
-    def generate_and_export_data_by_chunk(
-            context: SetupContext | GenIterContext,
-            stmt: GenerateStatement,
-            worker_id: int,
-            chunk_start: int,
-            chunk_end: int,
-            page_size: int,
-    ) -> dict:
-        """
-        Generate and export data by page in a single process.
-
-        :param context: SetupContext or GenIterContext instance.
-        :param stmt: GenerateStatement instance.
-        :param worker_id: Worker ID.
-        :param chunk_start: Start index of chunk.
-        :param chunk_end: End index of chunk.
-        :param page_size: Size of each page.
-        """
-
-        # Determine chunk data range, like (0, 1000), (1000, 2000), etc.
-        index_chunk = [(i, min(i + page_size, chunk_end)) for i in range(chunk_start, chunk_end, page_size)]
-
-        # Check if product result should be returned for test mode or memstore exporter
-        return_product_result = context.root.test_mode or any(
-            [context.root.memstore_manager.contain(exporter_str) for exporter_str in stmt.targets]
-        )
-        result: dict = {}
-
-        # Initialize ARTIFACT exporter state manager for each worker
-        exporter_state_manager = ExporterStateManager(worker_id)
-
-        # Create and cache exporters for each worker
-        exporters_set = stmt.targets.copy()
-        root_context = context.root
-
-        # Create exporters with operations
-        (
-            consumers_with_operation,
-            consumers_without_operation,
-        ) = root_context.class_factory_util.get_exporter_util().create_exporter_list(
-            setup_context=root_context,
-            stmt=stmt,
-            targets=list(exporters_set),
-        )
-
-        # Cache the exporters
-        root_context.task_exporters[stmt.full_name] = {
-            "with_operation": consumers_with_operation,
-            "without_operation": consumers_without_operation,
-            "page_count": 0,  # Track number of pages processed
-        }
-
-        # Generate and consume product by page
-        for index_tuple in index_chunk:
-            page_start, page_end = index_tuple
-            # Generate product
-            result_dict = generate_product_by_page_in_single_process(context, stmt, page_start, page_end, worker_id)
-
-            # Export product by page
-            context.root.class_factory_util.get_task_util_cls().export_product_by_page(
-                context.root, stmt, result_dict, exporter_state_manager
-            )
-
-            # Collect result for later capturing
-            if return_product_result:
-                for key, value in result_dict.items():
-                    result[key] = result.get(key, []) + value
-
-        # Check if product result should be returned for test mode or memstore exporter
-        has_memstore_exporter = any(
-            [
-                ("." not in exporter_str)
-                and ("(" not in exporter_str)
-                and context.root.memstore_manager.contain(exporter_str)
-                for exporter_str in stmt.targets
-            ]
-        )
-
-        # Return results if inner gen_stmt or test mode or memstore exporter
-        if isinstance(context, GenIterContext) or context.root.test_mode or has_memstore_exporter:
-            return result
-
-        return {}
-
-
-class RayGenerateWorker(GenerateWorker):
-    """
-    Worker class for generating and exporting data by page in multiprocessing using Ray.
-    """
-
-    @staticmethod
-    @ray.remote
-    def generate_and_export_data_by_chunk_mp(
-            context: SetupContext | GenIterContext,
-            stmt: GenerateStatement,
-            worker_id: int,
-            chunk_start: int,
-            chunk_end: int,
-            page_size: int,
-    ) -> dict:
-        """
-        Ray remote function to generate and export data by page in multiprocessing.
-        """
-        # Deserialize multiprocessing arguments
-        context.root.namespace.update(dill.loads(context.root.namespace_functions))
-        context.root.generators = dill.loads(context.root.generators)
-
-        return RayGenerateWorker.generate_and_export_data_by_chunk(context, stmt, worker_id, chunk_start, chunk_end,
-                                                                   page_size)
