+++ tests_ce/integration_tests/data_iteration/test_simple_iterate_json.xml
@@ -1,5 +1,5 @@
-<setup  multiprocessing="True" numProcess="2">
-    <generate name="product" source="data/people.json" count="1000" target="" cyclic="True">
+<setup numProcess="2">
+    <generate name="product" source="data/people.json" count="100000" target="" cyclic="True">
         <key name="addition" values="'Bob','Frank','Phil'"/>
         <key name="name" values="'Bob','Frank','Phil'"/>
     </generate>
